+++ src/template.rs
@@ -1,6 +1,7 @@
 use crate::command::{
     CompilationCommand, HookCommand, HookCommands, JudgingCommand, TranspilationCommand,
 };
+use crate::config;
 use crate::errors::{ExpandTemplateErrorKind, ExpandTemplateResult, StdError};
 use crate::path::{AbsPath, AbsPathBuf};
 use crate::service::ServiceKind;
@@ -90,11 +91,8 @@ impl<T: Target> Template<T> {
         self
     }
 
-    pub(crate) fn envs(mut self, envs: Option<&HashMap<String, String>>) -> Self {
-        if let Some(envs) = envs {
-            self.envs
-                .extend(envs.iter().map(|(k, v)| (k.clone(), v.clone())));
-        }
+    pub(crate) fn envs(mut self, envs: HashMap<String, String>) -> Self {
+        self.envs.extend(envs);
         self
     }
 }
@@ -116,13 +114,15 @@ impl Template<AbsPathBuf> {
             base_dir,
             service,
             contest,
+            mode,
         } = &self.requirements;
-        let envs = setup_env_vars(problem, &self.envs, *service, contest);
+        let envs = setup_env_vars(problem, &self.envs, *service, contest, *mode);
         let vars = {
             let mut vars = hashmap!(
                 "service" => OsString::from(service.to_string()),
                 "contest" => OsString::from(contest),
             );
+            vars.extend(mode.map(|m| ("mode", m.to_string().into())));
             vars.extend(problem.map(|p| ("problem", p.into())));
             vars.extend(self.extension.map(|e| ("extension", e.to_string().into())));
             vars
@@ -189,14 +189,15 @@ impl Template<TranspilationCommand> {
             base_dir,
             service,
             contest,
+            mode,
             shell,
             working_dir: TemplateBuilder(working_dir),
             src: TemplateBuilder(src),
             transpiled,
         } = &self.requirements;
 
-        let mut vars = setup_template_vars(Some(problem), *service, contest);
-        let mut envs = setup_env_vars(Some(problem), &self.envs, *service, contest);
+        let mut vars = setup_template_vars(Some(problem), *service, contest, *mode);
+        let mut envs = setup_env_vars(Some(problem), &self.envs, *service, contest, Some(*mode));
 
         let wd = working_dir.expand_as_path(base_dir, &vars, &envs)?;
         let src = src.expand_as_path(base_dir, &vars, &envs)?;
@@ -239,6 +240,7 @@ impl Template<CompilationCommand> {
             base_dir,
             service,
             contest,
+            mode,
             shell,
             working_dir: TemplateBuilder(working_dir),
             src: TemplateBuilder(src),
@@ -246,8 +248,8 @@ impl Template<CompilationCommand> {
             bin,
         } = &self.requirements;
 
-        let mut vars = setup_template_vars(Some(problem), *service, contest);
-        let mut envs = setup_env_vars(Some(problem), &self.envs, *service, contest);
+        let mut vars = setup_template_vars(Some(problem), *service, contest, *mode);
+        let mut envs = setup_env_vars(Some(problem), &self.envs, *service, contest, Some(*mode));
 
         let wd = working_dir.expand_as_path(base_dir, &vars, &envs)?;
         let src = src.expand_as_path(base_dir, &vars, &envs)?;
@@ -299,6 +301,7 @@ impl Template<JudgingCommand> {
             base_dir,
             service,
             contest,
+            mode,
             shell,
             working_dir: TemplateBuilder(working_dir),
             src: TemplateBuilder(src),
@@ -307,8 +310,8 @@ impl Template<JudgingCommand> {
             crlf_to_lf,
         } = &self.requirements;
 
-        let mut vars = setup_template_vars(Some(problem), *service, contest);
-        let mut envs = setup_env_vars(Some(problem), &self.envs, *service, contest);
+        let mut vars = setup_template_vars(Some(problem), *service, contest, *mode);
+        let mut envs = setup_env_vars(Some(problem), &self.envs, *service, contest, Some(*mode));
 
         let wd = working_dir.expand_as_path(base_dir, &vars, &envs)?;
         let src = src.expand_as_path(base_dir, &vars, &envs)?;
@@ -357,10 +360,12 @@ fn setup_template_vars(
     problem: Option<&str>,
     service: ServiceKind,
     contest: &str,
+    mode: config::Mode,
 ) -> HashMap<&'static str, OsString> {
     let mut ret = hashmap!(
         "service" => OsString::from(service.to_string()),
         "contest" => OsString::from(contest),
+        "mode" => OsString::from(mode.to_string()),
     );
     ret.extend(problem.map(|p| ("problem", p.into())));
     ret
@@ -371,6 +376,7 @@ fn setup_env_vars(
     utf8_envs: &HashMap<String, String>,
     service: ServiceKind,
     contest: &str,
+    mode: Option<config::Mode>,
 ) -> HashMap<String, OsString> {
     let mut ret = problem
         .map::<HashMap<String, OsString>, _>(|problem| {
@@ -393,6 +399,7 @@ fn setup_env_vars(
             )
         })
         .unwrap_or_default();
+    ret.extend(mode.map(|m| ("SNOWCHAINS_MODE".into(), m.to_string().into())));
     for (name, value) in utf8_envs {
         ret.insert(name.as_str().into(), value.as_str().into());
     }
@@ -462,6 +469,7 @@ pub(crate) struct AbsPathBufRequirements {
     pub(crate) base_dir: AbsPathBuf,
     pub(crate) service: ServiceKind,
     pub(crate) contest: String,
+    pub(crate) mode: Option<config::Mode>,
 }
 
 #[derive(Clone)]
@@ -476,6 +484,7 @@ pub(crate) struct TranspilationCommandRequirements {
     pub(crate) base_dir: AbsPathBuf,
     pub(crate) service: ServiceKind,
     pub(crate) contest: String,
+    pub(crate) mode: config::Mode,
     pub(crate) shell: HashMap<String, Vec<TemplateBuilder<OsString>>>,
     pub(crate) working_dir: TemplateBuilder<AbsPathBuf>,
     pub(crate) src: TemplateBuilder<AbsPathBuf>,
@@ -487,6 +496,7 @@ pub(crate) struct CompilationCommandRequirements {
     pub(crate) base_dir: AbsPathBuf,
     pub(crate) service: ServiceKind,
     pub(crate) contest: String,
+    pub(crate) mode: config::Mode,
     pub(crate) shell: HashMap<String, Vec<TemplateBuilder<OsString>>>,
     pub(crate) working_dir: TemplateBuilder<AbsPathBuf>,
     pub(crate) src: TemplateBuilder<AbsPathBuf>,
@@ -499,6 +509,7 @@ pub(crate) struct JudgingCommandRequirements {
     pub(crate) base_dir: AbsPathBuf,
     pub(crate) service: ServiceKind,
     pub(crate) contest: String,
+    pub(crate) mode: config::Mode,
     pub(crate) shell: HashMap<String, Vec<TemplateBuilder<OsString>>>,
     pub(crate) working_dir: TemplateBuilder<AbsPathBuf>,
     pub(crate) src: TemplateBuilder<AbsPathBuf>,
@@ -945,6 +956,7 @@ mod tests {
                     base_dir: base_dir(),
                     service: ServiceKind::Atcoder,
                     contest: "arc100".to_owned(),
+                    mode: None,
                 })
                 .expand(Some("ProblemName"))
                 .map_err(Into::into)
@@ -979,6 +991,7 @@ mod tests {
                         base_dir: base_dir(),
                         service: ServiceKind::Atcoder,
                         contest: "arc100".to_owned(),
+                        mode: None,
                     })
                     .expand(None)
                     .map_err(Into::into)